+++ packages/shared/src/types/database.types.ts
@@ -251,6 +251,7 @@ export type Database = {
           id: string
           last_active: string
           messages: Json
+          state: Json
           user_id: string
         }
         Insert: {
@@ -260,6 +261,7 @@ export type Database = {
           id?: string
           last_active?: string
           messages?: Json
+          state?: Json
           user_id: string
         }
         Update: {
@@ -269,6 +271,7 @@ export type Database = {
           id?: string
           last_active?: string
           messages?: Json
+          state?: Json
           user_id?: string
         }
         Relationships: [